Renamed and refractored commands and files names
