+++ client/src/app/utils/api-playground-json-parser.ts
