+++ src/app.ts
@@ -7,10 +7,10 @@ import routes from '@routes/routes';
 import util from 'util';
 import { exec } from 'child_process';
 import * as dotenv from 'dotenv';
+import curlCommands from '@utils/curl-commands-utils'
 
 dotenv.config();
 const execute = util.promisify(exec);
-const SIFS_API_KEY = process.env.SIFS_API_KEY;
 
 const app = express();
 app.use(bodyParser.json());
@@ -39,24 +39,13 @@ app.get('/', (_, res) => {
 });
 
 app.use('/api/v1/users', users);
+
 app.use('/api', routes);
 app.use(express.json());
-app.post('/api/test', async (req, res, next) => {
-  try {
-    const response = await execute(
-      `curl --connect-to "::gfe-wk.l.google.com" --header "Sec-Google-GFE-Restrict-Backends: shared-layer2-gfe=wk" "https://geoservicesinteractions-pa.googleapis.com/v1/availability:getAvailablePlaygroundApis?key=${SIFS_API_KEY}"`
-    );
-    res.json(JSON.parse(response.stdout));
-  } catch (e) {
-    next(e);
-  }
-});
 
 app.get('/api/ping', async (req, res, next) => {
   try {
-    const response = await execute(
-      `curl --connect-to "::gfe-wk.l.google.com" --header "Sec-Google-GFE-Restrict-Backends: shared-layer2-gfe=wk" "https://geoservicesinteractions-pa.googleapis.com/v1/ping:pingServer?key=${SIFS_API_KEY}"`
-    );
+    const response = await execute(curlCommands.PING_SERVER_API_COMMAND);
     res.json(JSON.parse(response.stdout));
   } catch (e) {
     next(e);
