+++ src/bin/www.ts
@@ -31,7 +31,7 @@ server.on('error', onError);
 server.on('listening', onListening);
 
 /**
- * Normalize a port into a number, string, or false.
+ * This method normalizes a port into a number, string, or false.
  */
 
 function normalizePort(val) {
